+++ src/structure/material.rs
@@ -6,13 +6,13 @@ pub struct Material {
     pub color: Color,
     pub pattern: Option<Pattern>,
 
-    pub shininess: f32,  // between 10 and 200
-    pub ambient: f32,    // between 0 and 1
-    pub diffuse: f32,    // between 0 and 1
-    pub specular: f32,   // between 0 and 1
-    pub reflective: f32, // between 0 and 1
-    pub refractive_index: f32,
-    pub transparency: f32,
+    pub shininess: f64,  // between 10 and 200
+    pub ambient: f64,    // between 0 and 1
+    pub diffuse: f64,    // between 0 and 1
+    pub specular: f64,   // between 0 and 1
+    pub reflective: f64, // between 0 and 1
+    pub refractive_index: f64,
+    pub transparency: f64,
 }
 
 impl Default for Material {
@@ -52,31 +52,31 @@ impl Material {
         self
     }
 
-    pub fn reflective(mut self, reflective: f32) -> Self {
+    pub fn reflective(mut self, reflective: f64) -> Self {
         self.reflective = reflective.clamp(0.0, 1.0);
 
         self
     }
 
-    pub fn specular(mut self, specular: f32) -> Self {
+    pub fn specular(mut self, specular: f64) -> Self {
         self.specular = specular.clamp(0.0, 1.0);
 
         self
     }
 
-    pub fn diffuse(mut self, diffuse: f32) -> Self {
+    pub fn diffuse(mut self, diffuse: f64) -> Self {
         self.diffuse = diffuse.clamp(0.0, 1.0);
 
         self
     }
 
-    pub fn shininess(mut self, shininess: f32) -> Self {
+    pub fn shininess(mut self, shininess: f64) -> Self {
         self.shininess = shininess.clamp(10.0, 200.0);
 
         self
     }
 
-    pub fn ambient(mut self, ambient: f32) -> Self {
+    pub fn ambient(mut self, ambient: f64) -> Self {
         self.ambient = ambient.clamp(0.0, 1.0);
 
         self
