+++ src/structure/world.rs
@@ -32,7 +32,7 @@ impl World {
         self.lights.push(light);
     }
 
-    pub fn intersect(&self, ray: Ray, n1: f32) -> Option<Intersection> {
+    pub fn intersect(&self, ray: Ray, n1: f64) -> Option<Intersection> {
         let mut closest_intersection = None;
 
         for object in &self.objects {
