+++ src/traits/transform.rs
@@ -1,7 +1,7 @@
 use crate::Vector;
 
 pub trait Transform {
-    fn rotate(&mut self, pitch: f32, yaw: f32, roll: f32);
+    fn rotate(&mut self, pitch: f64, yaw: f64, roll: f64);
     fn translate(&mut self, vec: Vector);
     fn scale(&mut self, vec: Vector);
 }
