+++ src/utils.rs
@@ -1,13 +1,13 @@
 use crate::constants::EPSILON;
 use std::ops::{Add, Mul, Sub};
 
-pub fn are_almost_equal(x: f32, y: f32) -> bool {
+pub fn are_almost_equal(x: f64, y: f64) -> bool {
     (x - y).abs() < EPSILON
 }
 
-pub fn lerp<T>(start: T, end: T, t: f32) -> T
+pub fn lerp<T>(start: T, end: T, t: f64) -> T
 where
-    T: Sub<Output = T> + Add<Output = T> + Mul<f32, Output = T>,
+    T: Sub<Output = T> + Add<Output = T> + Mul<f64, Output = T>,
 {
     let t = t.clamp(0., 1.);
